HTML5: Google rel="publisher" shall be link tag
Actual Google rel="publisher" is generated as <meta> tag. This causes
errors in HTML5 validation because <meta> has no attribute "rel" and
also, <meta> shall have a name, itemprop, property or http-equive
attributes.

Changing it to <link> tag solves these validations errors and makes
Google rich snippets work. It is important to notie that rel="publisher"
is meant to link a Google+ business page and not to individuals G+ page.
I don´t know if this affects or not the result but for future we shall
think in have a 'business' property in theme _config.yml and author´s as
already set. Then we might create rel="publisher" pointing to the
business and rel="author" linking to author´s G+ page.

Signed-off-by: Josenivaldo Benito Jr <jrbenito@benito.qsl.br>